Move parser to dedicated file
