+++ src/main/kotlin/fr/o80/release/parser/Parser.kt
@@ -1,0 +1,7 @@
+package fr.o80.release.parser
+
+import java.io.File
+
+interface Parser {
+    fun parse(file: File): ParsedFile
+}
